Refactor ether-cat-soft into csuethercat to make into a pip-installable package independent of the LRIS2 CSU.
