--- old_csu_sw/.gitignore
@@ -1,3 +1,0 @@
-
-/.idea
-/.venv
