--- old_csu_sw/_Archive/ueye_camera_parameters.ini
@@ -1,210 +1,0 @@
-﻿[Versions]
-uEye_api_64.dll=4.96.3985
-uEye_usb_64.sys=4.96.1328
-
-
-[Sensor]
-Sensor=UI149xLE-M
-Sensor bit depth=0
-Sensor source gain=0
-FPN correction mode=0
-Black reference mode=0
-Sensor digital gain=0
-
-
-[Image size]
-Start X=0
-Start Y=0
-Start X absolute=1
-Start Y absolute=1
-Width=3840
-Height=2748
-Binning=0
-Subsampling=0
-
-
-[Scaler]
-Mode=0
-Factor=1.000000
-
-
-[Multi AOI]
-Enabled=0
-Mode=0
-x1=0
-x2=0
-x3=0
-x4=0
-y1=0
-y2=0
-y3=0
-y4=0
-
-
-[Shutter]
-Mode=1
-Linescan number=0
-
-
-[Log Mode]
-Mode=3
-Manual value=0
-Manual gain=0
-
-
-[Timing]
-Pixelclock=30
-Extended pixelclock range=0
-Framerate=2.004174
-Exposure=267.712267
-Long exposure=0
-Dual exposure ratio=0
-
-
-[Selected Converter]
-IS_SET_CM_RGB32=1
-IS_SET_CM_RGB24=1
-IS_SET_CM_RGB16=1
-IS_SET_CM_RGB15=1
-IS_SET_CM_Y8=1
-IS_SET_CM_RGB8=1
-IS_SET_CM_BAYER=8
-IS_SET_CM_UYVY=1
-IS_SET_CM_UYVY_MONO=1
-IS_SET_CM_UYVY_BAYER=1
-IS_CM_CBYCRY_PACKED=0
-IS_SET_CM_RGBY=8
-IS_SET_CM_RGB30=8
-IS_SET_CM_Y12=8
-IS_SET_CM_BAYER12=8
-IS_SET_CM_Y16=8
-IS_SET_CM_BAYER16=8
-IS_CM_BGR12_UNPACKED=8
-IS_CM_BGRA12_UNPACKED=8
-IS_CM_JPEG=0
-IS_CM_SENSOR_RAW10=8
-IS_CM_MONO10=8
-IS_CM_BGR10_UNPACKED=8
-IS_CM_RGBA8_PACKED=1
-IS_CM_RGB8_PACKED=1
-IS_CM_RGBY8_PACKED=8
-IS_CM_RGB10V2_PACKED=8
-IS_CM_RGB12_UNPACKED=8
-IS_CM_RGBA12_UNPACKED=8
-IS_CM_RGB10_UNPACKED=8
-IS_CM_RGB8_PLANAR=1
-
-
-[Parameters]
-Colormode=11
-Gamma=1.000000
-Hardware Gamma=0
-Blacklevel Mode=1
-Blacklevel Offset=0
-Hotpixel Mode=2
-Hotpixel Threshold=0
-Sensor Hotpixel=0
-Adaptive hotpixel correction enable=0
-Adaptive hotpixel correction mode=0
-Adaptive hotpixel correction sensitivity=3
-GlobalShutter=0
-AllowRawWithLut=0
-
-
-[Gain]
-Master=0
-Red=0
-Green=0
-Blue=0
-GainBoost=0
-
-
-[Processing]
-EdgeEnhancementFactor=0
-RopEffect=0
-Whitebalance=0
-Whitebalance Red=1.000000
-Whitebalance Green=1.000000
-Whitebalance Blue=1.000000
-Color correction=0
-Color_correction_factor=1.000000
-Color_correction_satU=100
-Color_correction_satV=100
-Bayer Conversion=1
-JpegCompression=0
-NoiseMode=0
-ImageEffect=0
-LscModel=0
-WideDynamicRange=0
-
-
-[Auto features]
-Auto Framerate control=0
-Brightness exposure control=0
-Brightness gain control=0
-Auto Framerate Sensor control=0
-Brightness exposure Sensor control=0
-Brightness gain Sensor control=0
-Brightness exposure Sensor control photometry=0
-Brightness gain Sensor control photometry=0
-Brightness control once=0
-Brightness reference=128
-Brightness speed=50
-Brightness max gain=100
-Brightness max exposure=498.848500
-Brightness Aoi Left=0
-Brightness Aoi Top=0
-Brightness Aoi Width=3840
-Brightness Aoi Height=2748
-Brightness Hysteresis=2
-AutoImageControlMode=2
-AutoImageControlPeakWhiteChannel=0
-AutoImageControlExposureMinimum=0.000000
-AutoImageControlPeakWhiteChannelMode=0
-AutoImageControlPeakWhiteGranularity=0
-Auto WB control=0
-Auto WB type=2
-Auto WB RGB color model=1
-Auto WB RGB color temperature=0
-Auto WB offsetR=0
-Auto WB offsetB=0
-Auto WB gainMin=0
-Auto WB gainMax=100
-Auto WB speed=50
-Auto WB Aoi Left=0
-Auto WB Aoi Top=0
-Auto WB Aoi Width=3840
-Auto WB Aoi Height=2748
-Auto WB Once=0
-Auto WB Hysteresis=2
-Brightness Skip Frames Trigger Mode=4
-Brightness Skip Frames Freerun Mode=4
-Auto WB Skip Frames Trigger Mode=4
-Auto WB Skip Frames Freerun Mode=4
-
-
-[Trigger and Flash]
-Trigger mode=0
-Trigger timeout=200
-Trigger delay=0
-Trigger debounce mode=0
-Trigger debounce delay time=0
-Trigger burst size=1
-Trigger prescaler frame=1
-Trigger prescaler line=1
-Trigger input=1
-Flash strobe=0
-Flash delay=0
-Flash duration=0
-Flash auto freerun=0
-PWM mode=0
-PWM frequency=20000000
-PWM dutycycle=20000000
-GPIO state=0
-GPIO direction=3
-GPIO1 Config=2
-GPIO2 Config=2
-
-
-[Memory]
-Camera memory mode=0
