--- old_csu_sw/uEye_SimpleLive_PyuEye_OpenCV/doc/README_Linux
@@ -1,50 +1,0 @@
-Created on Ubuntu 16.04.3 LTS - Instructions for Python and Python3 under Linux
-
-
-
-
-
-Open a Terminal: [Ctrl]+[Alt]+[T]
-
-
-
-Check if Python is already installed: python -V
-
-Check if Python3 is already installed: python3 -V
-
-
-
-If there is no Python, install it: sudo apt-get install python
-
-If there is no Python3, install it: sudo apt-get install python3
-
-
-
-Install pip for Python: sudo apt-get install python-pip
-
-Install pip for Python3: sudo apt-get install python3-pip
-
-
-
-Now install the ueye wrapper: pip install pyueye
-
-
-
-(Additional you may need some OpenCV libraries to display the image. In that case please search in the internet for guids to install OpenCV.)
-
-Move to your directory containing the Python Demo, right click and select "open in terminal".
-
-Then enter the following: python SimpleLive_Python_uEye_OpenCV.py
-
-This should execute the code and give you a live-video.
-
-
-
-To access the code, simple double-left click on the .py document
-
-
-
-
-
-Address to wrapper Description + Quick Start and License: 
-https://pypi.org/project/pyueye/
