--- old_csu_sw/uEye_SimpleLive_PyuEye_OpenCV/doc/README_Windows.txt
@@ -1,21 +1,0 @@
-Created on Windows 7 Professional - Instructions for Python under Windows
-
-Good Tutorial:
-https://www.solarianprogrammer.com/2016/09/17/install-opencv-3-with-python-3-on-windows/
-
-
-
-Go to the official Python website: https://www.python.org/
-Download and install your preferable Python Version in den "Downloads" section.
-When installing do enable additional "Precompile Standard library" and "Add Python 3.6 to PATH".
-
-You will also need numpy and OpenCV. I advise to install them trough their wheel in the console.
-For example, I downloaded and installed numpy with: pip install numpy-1.14.1+mkl-cp36-cp36m-win_amd64.whl
-And OpenCV with: pip install opencv_python-3.4.1-cp36-cp36m-win_amd64.whl
-
-You will also need the uEye library.
-Go to the official IDS website under the "Download" section, select "Interfaces" and then "Python".
-Read the description and click on the link: https://pypi.org/project/pyueye/
-
-You can download the needed package in the "Download Files" tab. 
-Install it once more trough the console with: pip install pyueye-4.90.0.0-py2.py3-none-any.whl
