--- run_hom.sh
@@ -1,9 +1,0 @@
-#!/bin/bash
-
-HOST="131.215.193.41"
-PORT=7469
-
-# Send Homing PDO message
-python3 commands.py $HOST $PORT --send-homingpdo
-
-
