--- settings/masterSettings.yaml
@@ -1,1 +1,0 @@
-networkInterfaceName: enx6c6e071121bf
