--- settings/serverSettings.yaml
@@ -1,2 +1,0 @@
-host: 131.215.193.25
-port: 7469
