--- src/HAL/pysoem/pysoemHelpers.py
@@ -1,5 +1,0 @@
-import pysoem
-import struct
-
-from enum import Enum
-
