--- test/sample_positions.csv
@@ -1,3 +1,0 @@
-1000,2000,3000,10
-1500,2500,3500,15
-2000,3000,4000,20
